+++ R/delim.R
@@ -1,13 +1,23 @@
 #' A simply write txt to a temp file with a logging feature
-#' @return path to temp txt file
-#' @param x data to write
-#' @param delimiter delimiter for the readr::write_delim argument
-#' @param ... additional arguments for the readr::write_delim argument.
+#' @description
+#' Write a delimited text file to a temporary path. Path to the temporary text file is returned.
 #' @importFrom readr write_delim
+#' @inheritParams readr::write_delim
+#' @rdname write_temp_delim
 #' @export
+#' @importFrom readr write_delim
+#' @family delim functions
 
 write_temp_delim <-
-        function(x, delimiter, log = TRUE, log_details = "", ...) {
+        function(x,
+                 delim = " ",
+                 na = NA,
+                 append = FALSE,
+                 col_name = !append,
+                 quote_escape = double,
+                 eol = "\n",
+                 log = "DEPRECATED",
+                 log_details = "") {
 
 
                 temp_file <- tempfile(fileext = ".txt")
@@ -19,10 +29,16 @@ write_temp_delim <-
                                  details = log_details)
                 }
 
-                readr::write_delim(x = x,
-                                 path = temp_file,
-                                 delim = delimiter,
-                                 ...)
+                readr::write_delim(
+                        x = x,
+                        file = temp_file,
+                        delim = delim,
+                        na = na,
+                        append = append,
+                        col_names = col_names,
+                        quote_escape = quote_escape,
+                        eol = eol
+                )
 
                 return(temp_file)
         }